refactor: rename file to utils.tsx
Renames file to utils.tsx as suggested in PR review.
