+++ bigbluebutton-html5/imports/ui/components/plugins/plugin-icon/utils.tsx
